Use FileEntry ptrs instead of paths for tracking included files
In the past, the same instance of Inliner was used for multiple
compilation units. This required tracking via file paths, because the
same file would be represented by different FileEntry objects during
processing different compilation units. Now that we concatenate all
compilation units in advance, this is not necessary anymore.